refactor(BA-6611): move repository-consumed specs/exceptions out of the service layer
Remove repositories/* -> services upward import edges in the manager component.

- Relocate UserCreateSpec/UserUpdateSpec from services/user/types.py to the
  repository layer, following the per-domain convention (specs live in
  creators.py/updaters.py; types.py is SearchScope-only). Delete the now-empty
  services/user/types.py and its unused NoUserUpdateError.
- Consolidate model_serving exceptions onto the canonical errors/ definitions
  (errors.common, errors.api, errors.service), add the missing
  EndpointAutoScalingRuleNotFound to errors/service.py, and delete the duplicate
  services/model_serving/exceptions.py.
