+++ tests/unit/manager/services/model_serving/actions/test_delete_auto_scaling_rule.py
@@ -12,14 +12,14 @@ from ai.backend.common.data.user.types import UserData, UserRole
 from ai.backend.common.types import RuleId
 from ai.backend.manager.actions.monitors.monitor import ActionMonitor
 from ai.backend.manager.actions.validators import ActionValidators
+from ai.backend.manager.errors.service import (
+    EndpointAutoScalingRuleNotFound,
+)
 from ai.backend.manager.repositories.model_serving.repositories import ModelServingRepositories
 from ai.backend.manager.repositories.model_serving.repository import ModelServingRepository
 from ai.backend.manager.services.model_serving.actions.delete_auto_scaling_rule import (
     DeleteEndpointAutoScalingRuleAction,
     DeleteEndpointAutoScalingRuleActionResult,
-)
-from ai.backend.manager.services.model_serving.exceptions import (
-    EndpointAutoScalingRuleNotFound,
 )
 from ai.backend.manager.services.model_serving.processors.auto_scaling import (
     ModelServingAutoScalingProcessors,
